StoikVisnyk/dailyMotivationApp/daily_message/requirements.txt
anyio==4.4.0
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==1.26.19
